rename, move to common file
