+++ src/components/Summary.jsx
@@ -7,7 +7,7 @@ import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
 import { QUESTIONNAIRE_ANCHOR_ID_PREFIX } from "../consts/consts";
 import Questionnaire from "../models/Questionnaire";
-import { hasData, isEmptyArray } from "../util/util";
+import { hasData, isEmptyArray } from "../util";
 import Error from "./ErrorComponent";
 import QuestionnaireInfo from "./QuestionnaireInfo";
 import Responses from "./Responses";
