+++ src/components/TimeoutModal.jsx
@@ -2,7 +2,7 @@ import React, {useState} from "react";
 import Box from '@mui/material/Box';
 import Modal from '@mui/material/Modal';
 import Timeout from "../util/timeout";
-import { getEnvDashboardURL } from "../util/util";
+import { getEnvDashboardURL } from "../util";
 
 const style = {
   position: 'absolute',
