+++ src/components/Version.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Box from "@mui/material/Box";
-import { getEnv } from "../util/util";
+import { getEnv } from "../util";
 export default function Version() {
   const versionString = getEnv("REACT_APP_VERSION_STRING");
   if (!versionString) return null;
