+++ src/containers/LaunchContainer.jsx
@@ -5,7 +5,7 @@ import CircularProgress from "@mui/material/CircularProgress";
 import { ThemeProvider } from "@mui/material/styles";
 import ErrorComponent from "../components/ErrorComponent";
 import { queryNeedPatientBanner, queryPatientIdKey } from "../consts/consts";
-import { fetchEnvData, getEnv } from "../util/util";
+import { fetchEnvData, getEnv } from "../util";
 import { getTheme } from "../config/theme_config";
 import "../style/App.scss";
